+++ lib/SNGBEcomm.php
@@ -16,6 +16,7 @@ if (!function_exists('curl_init')) {
 require(dirname(__FILE__) . '/SNGBEcomm/SNGBEcomm.php');
 
 require(dirname(__FILE__) . '/SNGBEcomm/Payment.php');
+require(dirname(__FILE__) . '/SNGBEcomm/Error.php');
 // Utilities
 //require(dirname(__FILE__) . '/SNGBEcomm/Util.php');
 //require(dirname(__FILE__) . '/SNGBEcomm/Util/Set.php');