+++ CHANGELOG.md
@@ -1,8 +1,15 @@
+0.1.3
+-----
+* Create basic payment behavior ([@zatsepin](mailto:yu.zatsepin@sngb.ru))
+* Create sampleshop ([@zatsepin](mailto:yu.zatsepin@sngb.ru))
+
 0.1.2
 -----
 * Create init request and get result url and payment id ([@zatsepin](mailto:yu.zatsepin@sngb.ru))
 
-
 0.1.1
 -----
 * Initial git for VC ([@zatsepin](mailto:yu.zatsepin@sngb.ru))
+
+
+